+++ apis/keda/v1alpha1/scaledobject_webhook.go
@@ -36,7 +36,6 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/webhook"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
 	metricscollector "github.com/kedacore/keda/v2/pkg/metricscollector/webhook"
@@ -58,9 +57,8 @@ func (so *ScaledObject) SetupWebhookWithManager(mgr ctrl.Manager, cacheMissFallb
 		return fmt.Errorf("failed to setup kubernetes clients: %w", err)
 	}
 
-	return ctrl.NewWebhookManagedBy(mgr).
+	return ctrl.NewWebhookManagedBy(mgr, &ScaledObject{}).
 		WithValidator(&ScaledObjectCustomValidator{}).
-		For(so).
 		Complete()
 }
 
@@ -91,35 +89,31 @@ func setupKubernetesClients(mgr ctrl.Manager, cacheMissFallback bool) error {
 // ScaledObjectCustomValidator is a custom validator for ScaledObject objects
 type ScaledObjectCustomValidator struct{}
 
-func (socv ScaledObjectCustomValidator) ValidateCreate(ctx context.Context, obj runtime.Object) (warnings admission.Warnings, err error) {
+func (socv ScaledObjectCustomValidator) ValidateCreate(ctx context.Context, so *ScaledObject) (warnings admission.Warnings, err error) {
 	request, err := admission.RequestFromContext(ctx)
 	if err != nil {
 		return nil, err
 	}
-	so := obj.(*ScaledObject)
 	return so.ValidateCreate(request.DryRun)
 }
 
-func (socv ScaledObjectCustomValidator) ValidateUpdate(ctx context.Context, oldObj, newObj runtime.Object) (warnings admission.Warnings, err error) {
+func (socv ScaledObjectCustomValidator) ValidateUpdate(ctx context.Context, old, so *ScaledObject) (warnings admission.Warnings, err error) {
 	request, err := admission.RequestFromContext(ctx)
 	if err != nil {
 		return nil, err
 	}
-	so := newObj.(*ScaledObject)
-	old := oldObj.(*ScaledObject)
 	return so.ValidateUpdate(old, request.DryRun)
 }
 
-func (socv ScaledObjectCustomValidator) ValidateDelete(ctx context.Context, obj runtime.Object) (warnings admission.Warnings, err error) {
+func (socv ScaledObjectCustomValidator) ValidateDelete(ctx context.Context, so *ScaledObject) (warnings admission.Warnings, err error) {
 	request, err := admission.RequestFromContext(ctx)
 	if err != nil {
 		return nil, err
 	}
-	so := obj.(*ScaledObject)
 	return so.ValidateDelete(request.DryRun)
 }
 
-var _ webhook.CustomValidator = &ScaledObjectCustomValidator{}
+var _ admission.Validator[*ScaledObject] = &ScaledObjectCustomValidator{}
 
 // ValidateCreate implements webhook.Validator so a webhook will be registered for the type
 func (so *ScaledObject) ValidateCreate(dryRun *bool) (admission.Warnings, error) {