+++ pkg/scalers/scalersconfig/typed_config.go
@@ -34,6 +34,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
+	"github.com/kedacore/keda/v2/pkg/common/action"
 	"github.com/kedacore/keda/v2/pkg/eventreason"
 )
 
@@ -253,7 +254,7 @@ func (sc *ScalerConfig) setValue(field reflect.Value, params Params) ([]string, 
 		if sc.Recorder != nil {
 			message := fmt.Sprintf("scaler %s info: %s", sc.TriggerType, params.DeprecatedAnnounce)
 			fmt.Print(message)
-			sc.Recorder.Event(sc.ScaledObject, corev1.EventTypeNormal, eventreason.KEDAScalersInfo, message)
+			sc.Recorder.Eventf(sc.ScaledObject, nil, corev1.EventTypeNormal, eventreason.KEDAScalersInfo, action.Info, message)
 		}
 	}
 	if !exists && params.Default != "" {
@@ -544,7 +545,7 @@ func (sc *ScalerConfig) checkUnexpectedParameterExist(parsedParamNames []string,
 				message := fmt.Sprintf("Unmatched input property %s in scaler %s", key+suffix, scalerIdentifier)
 				// Just logging as it's optional property checking and should not block the scaling
 				logger.Error(nil, message)
-				sc.Recorder.Event(sc.ScaledObject, corev1.EventTypeWarning, eventreason.KEDAScalersInfo, message)
+				sc.Recorder.Eventf(sc.ScaledObject, nil, corev1.EventTypeWarning, eventreason.KEDAScalersInfo, action.Info, message)
 			}
 		}
 	}