+++ pkg/scaling/cache/scalers_cache.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/expr-lang/expr/vm"
 	v2 "k8s.io/api/autoscaling/v2"
-	"k8s.io/client-go/tools/record"
+	"k8s.io/client-go/tools/events"
 	"k8s.io/metrics/pkg/apis/external_metrics"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
@@ -39,7 +39,7 @@ type ScalersCache struct {
 	ScaledObject             *kedav1alpha1.ScaledObject
 	Scalers                  []ScalerBuilder
 	ScalableObjectGeneration int64
-	Recorder                 record.EventRecorder
+	Recorder                 events.EventRecorder
 	CompiledFormula          *vm.Program
 	mutex                    sync.RWMutex
 }